+++ src/main/java/com/bitbank/entities/v1/UserEntity.java
@@ -1,0 +1,9 @@
+package com.bitbank.entities.v1;
+
+import lombok.Data;
+
+@Data
+public class UserEntity {
+	private String username;
+	private String password;
+}